util.c: parse_hex returns off_t instead of size_t

+++ src/internal.h
@@ -31,7 +31,7 @@ void stack_clean(NsonStack *stack);
 
 off_t parse_dec(int64_t *i, const char *p, size_t len);
 
-size_t parse_hex(uint64_t *dest, const char *src, size_t len);
+off_t parse_hex(uint64_t *dest, const char *src, size_t len);
 
 const char *parse_real(double *r, const char *p, size_t len);
